Alexander Dilemma Problem Solution
Alexendar’s Dilemma Problem - I owe my son Alexander $50 per month for doing chores around the house. Instead of giving him the $50, each month I will let him reach into a bag which contains a $100 bill and five $10 bills, and draw two of the bills. Should Alexander go with this scheme, or in other words, is my scheme fair? If Alexander is a risk taker, then he will have a chance of making $110 instead of $50 per month if he is lucky. On the other hand, if he is risk averse, then he figures that he will more likely be getting $20 rather than $50 per month. So the question is: In the long run, will he be better off taking the $50 or go with my scheme?
Solution for this problem is coded in python language